pyproject.toml
[tool.pytest.ini_options]
pythonpath = "."
[project]
name = "recipe-vision-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
